% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{browseFile}
\alias{browseFile}
\title{Browse a system file in the default browser}
\usage{
browseFile(url, browser = getOption("browser"), encodeIfNeeded = FALSE)
}
\arguments{
\item{url}{string: as for \code{\link[utils:browseURL]{utils::browseURL()}}}

\item{browser}{string: as for \code{\link[utils:browseURL]{utils::browseURL()}}}

\item{encodeIfNeeded}{logical: as for \code{\link[utils:browseURL]{utils::browseURL()}}}
}
\description{
RStudio overrides the default behaviour of browseURL on some platforms, meaning that local files are not opened as \verb{file:///...} URLs but as \verb{http://localhost...}. This can break some local HTML files that are expecting to be served as \verb{file:///} URLs.
}
\examples{
myfile <- tempfile(fileext = ".html")
cat("<h1>Hello!</h1>", file = myfile)

## in RStudio on Linux, this will be opened as a http://localhost URL
if (interactive()) browseURL(myfile)

## but this shouldn't
browseFile(myfile)

}
