% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{ov_cmat_estimate}
\alias{ov_cmat_estimate}
\title{Estimate the camera matrix}
\usage{
ov_cmat_estimate(X, x)
}
\arguments{
\item{X}{matrix or data.frame: Nx3 matrix of 3D real-world coordinates}

\item{x}{matrix or data.frame: Nx2 matrix of image coordinates}
}
\value{
A list with components \code{coef} (fitted transformation coefficients) and \code{rmse} (root mean squared error of the fitted transformation)
}
\description{
The camera matrix characterizes the mapping of a camera from 3D real-world coordinates to 2D coordinates in an image.
}
\examples{

## define real-world and corresponding image coordinates
xX <- dplyr::tribble(~image_x, ~image_y, ~court_x, ~court_y,   ~z,
                        0.054,    0.023,      0.5,      0.5,    0, ## near left baseline
                        0.951,    0.025,      3.5,      0.5,    0, ## near right baseline
                        0.752,    0.519,      3.5,      6.5,    0, ## far right baseline
                        0.288,    0.519,      0.5,      6.5,    0, ## far left baseline
                        0.199,    0.644,      0.5,      3.5, 2.43, ## left net top
                        0.208,    0.349,      0.5,      3.5, 0.00, ## left net floor
                        0.825,    0.644,      3.5,      3.5, 2.43, ## right net top
                        0.821,    0.349,      3.5,      3.5, 0.00) ## right net floor

C <- ov_cmat_estimate(X = xX[, 3:5], x = xX[, 1:2])

## fitted image coordinates using C
ov_cmat_apply(C, X = xX[, 3:5])

## compare to actual image positions
xX[, 1:2]

}
\references{
\url{https://en.wikipedia.org/wiki/Camera_matrix}. For general background see e.g. Ballard DH, Brown CM (1982) Computer Vision. Prentice-Hall, New Jersey
}
\seealso{
\code{\link[=ov_cmat_apply]{ov_cmat_apply()}}
}
