% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg_utils.R
\name{ov_create_video-deprecated}
\alias{ov_create_video-deprecated}
\title{Playlist to video file}
\usage{
ov_create_video(playlist, filename, subtitle_column = NULL)
}
\arguments{
\item{playlist}{data.frame: a playlist as returned by \code{ov_video_playlist}. Note that only local video sources are supported}

\item{filename}{string: file to write to. If not specified (or \code{NULL}), a file in the temporary directory will be created. If \code{filename} exists, it will be overwritten. The extension of \code{filename} will determine the output format}

\item{subtitle_column}{string: if not \code{NULL}, a subtitle file will be produced using the contents of this column (in the playlist) as the subtitle for each clip. The subtitle file will have the same name as \code{filename} but with extension ".srt"}
}
\value{
A list with the filenames of the created video and subtitle files.
}
\description{
Make a self-contained video file from a playlist
}
\seealso{
\code{\link{ovideo-deprecated}}
}
\keyword{internal}
