% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{ov_example_video}
\alias{ov_example_video}
\title{Example video clips provided as part of the ovideo package}
\usage{
ov_example_video(choice = 1)
}
\arguments{
\item{choice}{integer: which video file to return?
\itemize{
\item 1 - a clip from a match between GKS Katowice and MKS Bedzin during the 2018/19 Polish Plus Liga
}}
}
\value{
Path to the video file
}
\description{
Example video clips provided as part of the ovideo package
}
