% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg_utils.R
\name{ov_ffmpeg_exe}
\alias{ov_ffmpeg_exe}
\alias{ov_ffmpeg_ok}
\title{ffmpeg executable functions}
\usage{
ov_ffmpeg_exe()

ov_ffmpeg_ok(do_error = FALSE)
}
\arguments{
\item{do_error}{logical: if \code{TRUE}, throw an error if the ffmpeg executable cannot be found}
}
\value{
For \code{ov_ffmpeg_exe}, the path to the executable, or \code{NULL} if not found. For \code{ov_ffmpeg_ok}, a logical indicating whether the executable could be found or not
}
\description{
Helper functions to find the ffmpeg executable. If ffmpeg is not installed on the system, it can be installed (for some platforms) with \code{\link[=ov_install_ffmpeg]{ov_install_ffmpeg()}}.
}
\examples{
ov_ffmpeg_ok()

}
\seealso{
\code{\link[=ov_install_ffmpeg]{ov_install_ffmpeg()}}
}
