% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg_utils.R
\name{ov_find_video_file}
\alias{ov_find_video_file}
\title{Try and locate a video file, when the path embedded in the dvw file is for another computer}
\usage{
ov_find_video_file(dvw_filename, video_filename = NULL)
}
\arguments{
\item{dvw_filename}{string: the full path to the DataVolley file}

\item{video_filename}{character: one or more video file paths. If \code{NULL}, the video file name embedded in the DataVolley file will be used}
}
\value{
A character vector, with one entry per \code{video_filename}. Video files that could not be found will be \code{NA} here.
}
\description{
Try and locate a video file, when the path embedded in the dvw file is for another computer
}
