% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{ov_get_court_ref}
\alias{ov_get_court_ref}
\title{Define the reference points on a court image}
\usage{
ov_get_court_ref(image_file, video_file, t = 60, type = "corners")
}
\arguments{
\item{image_file}{string: path to an image file (jpg) containing the court image (not required if \code{video_file} is supplied)}

\item{video_file}{string: path to a video file from which to extract the court image (not required if \code{image_file} is supplied)}

\item{t}{numeric: the time of the video frame to use as the court image (not required if \code{image_file} is supplied)}

\item{type}{string: currently only "corners"}
}
\value{
A data.frame containing the reference information
}
\description{
This function is used to define the reference points on a court image, to be used with \code{\link[=ov_transform_points]{ov_transform_points()}}.
The court coordinate system is that used in \code{\link[datavolley:dv_court]{datavolley::dv_court()}}, \code{\link[datavolley:ggcourt]{datavolley::ggcourt()}}, and related functions.
Try \verb{plot(c(0, 4), c(0, 7), type = "n", asp = 1); datavolley::dv_court()} or \code{ggplot2::ggplot() + datavolley::ggcourt() + ggplot2::theme_bw()} for a visual depiction.
}
\examples{
if (interactive()) {
  crt <- ov_get_court_ref(image_file = system.file("extdata/2019_03_01-KATS-BEDS-court.jpg",
                          package = "ovideo"))

}
}
\seealso{
\code{\link[=ov_transform_points]{ov_transform_points()}}, \code{\link[datavolley:dv_court]{datavolley::dv_court()}},  \code{\link[datavolley:ggcourt]{datavolley::ggcourt()}}
}
