% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{ov_get_video_data}
\alias{ov_get_video_data}
\title{Retrieve a data object stored in a video file metadata tag}
\usage{
ov_get_video_data(video_file, tag = "ov_court_info", b64 = TRUE)
}
\arguments{
\item{video_file}{string: path to the video file}

\item{tag}{string: the tag name to use}

\item{b64}{logical: was \code{obj} serialized and base64-encoded before storing?}
}
\value{
The stored information, or \code{NULL} if there was none
}
\description{
Retrieve a data object stored in a video file metadata tag
}
\examples{
\dontrun{
  if (interactive()) {
    ## mark the geometry of the court in the video
    ref <- ov_shiny_court_ref(video_file = ov_example_video(), t = 5)

    ## store it
    newfile <- ov_set_video_data(ov_example_video(), obj = ref)

    ## retrieve it
    ov_get_video_data(newfile)
  }
}
}
\seealso{
\code{\link[=ov_set_video_data]{ov_set_video_data()}}
}
