% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg_utils.R
\name{ov_images_to_video}
\alias{ov_images_to_video}
\title{Encode a set of images into a video}
\usage{
ov_images_to_video(
  input_dir,
  image_file_mask = "image_\%06d.jpg",
  image_files,
  outfile,
  fps = 30,
  extra = NULL,
  debug = FALSE
)
}
\arguments{
\item{input_dir}{string: path to the input directory}

\item{image_file_mask}{string: the mask that specifies the image files, e.g. "image_\%06d.jpg" for images named "image_000001.jpg", "image_000002.jpg" etc}

\item{image_files}{character: vector of input image files, in order that they should appear in the video. Used if \code{input_dir} is missing}

\item{outfile}{string: the output file. If missing, a temporary file (with extension .mp4) will be used}

\item{fps}{numeric: frames per second}

\item{extra}{: additional parameters passed to ffmpeg, in the form c("param", "value", "param2", "value2"). For example, \code{c("-vb", "4096k")} could be used to control the output video bitrate}

\item{debug}{logical: if \code{TRUE}, echo the ffmpeg output to the console}
}
\value{
The path to the video file
}
\description{
Requires that ffmpeg is available on your system path. Input files can either be specified as a list of image files, or alternatively as a directory name and image file mask. For the latter, the images must be numbered in sequential order.
}
\seealso{
\code{\link[av:encoding]{av::av_encode_video()}} as an alternative
}
