% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{ov_merge_video_timing_df}
\alias{ov_merge_video_timing_df}
\title{Merge two video timing dataframes}
\usage{
ov_merge_video_timing_df(x, default = ov_video_timing_df())
}
\arguments{
\item{x}{data.frame: video timings to use}

\item{default}{data.frame: default timings to use, for anything not provided in \code{x}}
}
\value{
A data.frame
}
\description{
Merge two video timing dataframes
}
\examples{

my_timings <- data.frame(skill = "Attack", phase = "Reception", start_offset = 0)
ov_merge_video_timing_df(my_timings)

}
\seealso{
\code{\link[=ov_video_timing_df]{ov_video_timing_df()}}
}
