% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{ov_overlay_image}
\alias{ov_overlay_image}
\title{Generate a court overlay image showing court boundary, 3m, zone, and other lines}
\usage{
ov_overlay_image(court_ref, height, width, filename, ...)
}
\arguments{
\item{court_ref}{data.frame: as returned by \code{\link{ov_get_court_ref}}}

\item{height}{integer: height of image to produce in pixels}

\item{width}{integer: width of image to produce in pixels}

\item{filename}{string: image filename (png). If missing, a file will be created in the temporary directory}

\item{...}{: arguments passed to \code{\link{ov_overlay_data}}}
}
\value{
The path to the generated file.
}
\description{
Generate a court overlay image showing court boundary, 3m, zone, and other lines
}
