% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{ov_playlist_to_vlc}
\alias{ov_playlist_to_vlc}
\title{Convert playlist to VLC m3u format}
\usage{
ov_playlist_to_vlc(playlist, outfile, no_paths = FALSE, seamless = TRUE)
}
\arguments{
\item{playlist}{data.frame: as returned by \code{\link[=ov_video_playlist]{ov_video_playlist()}}}

\item{outfile}{string: the file name to write to. If not supplied, a file will be created in the temporary directory. Note that the directory of \code{outfile} must already exist}

\item{no_paths}{logical: if \code{TRUE}, remove the paths from video files. The m3u file must be saved into the same directory as the video source file(s)}

\item{seamless}{logical: if \code{TRUE}, merge adjacent items into a single clip}
}
\value{
The path to the m3u file
}
\description{
Converts a playlist object to a m3u file that can be opened with VLC. Note that the video files must be present on your local file system in order for this to work.
}
\references{
\url{https://www.videolan.org/}, \url{https://wiki.videolan.org/M3U/}
}
\seealso{
\code{\link[=ov_video_playlist]{ov_video_playlist()}}
}
