% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{ov_set_video_data}
\alias{ov_set_video_data}
\title{Store a data object in a video file metadata tag}
\usage{
ov_set_video_data(
  video_file,
  obj,
  tag = "ov_court_info",
  b64 = TRUE,
  replace = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{video_file}{string: path to the video file}

\item{obj}{: data object to store, typically a list as returned by \code{\link[=ov_shiny_court_ref]{ov_shiny_court_ref()}}. \code{obj} will be serialized and base64-encoded before storing unless \code{b64 = FALSE}}

\item{tag}{string: the tag name to use}

\item{b64}{logical: serialize \code{obj} and base64-encode before storing?}

\item{replace}{logical: if \code{FALSE} and the specified metadata tag is already present in the video file, don't replace it}

\item{overwrite}{logical: if \code{TRUE} overwrite the \code{video_file}, otherwise create a new file in the temporary directory. See \code{\link[=ov_set_video_meta]{ov_set_video_meta()}} for details}
}
\value{
The path to the video file
}
\description{
This function stores an R data object (data frame, list, etc) within a metadata tag inside a video file. This is primarily intended to store video-specific information, so that this information is carried with the video file itself. By default the \code{ov_court_info} metadata tag is used (intended to store the geometry of the playing court in the video, see Examples).
}
\examples{
\dontrun{
  if (interactive()) {
    ## mark the geometry of the court in the video
    ref <- ov_shiny_court_ref(video_file = ov_example_video(), t = 5)

    ## store it
    newfile <- ov_set_video_data(ov_example_video(), obj = ref)

    ## retrieve it
    ov_get_video_data(newfile)
  }
}

}
\seealso{
\code{\link[=ov_get_video_data]{ov_get_video_data()}}, \code{\link[=ov_set_video_meta]{ov_set_video_meta()}}
}
