% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{ov_set_video_meta}
\alias{ov_set_video_meta}
\title{Retrieve metadata tags from a video file}
\usage{
ov_set_video_meta(video_file, ..., overwrite = FALSE, debug = FALSE)
}
\arguments{
\item{video_file}{string: path to the video file}

\item{...}{: named values to set}

\item{overwrite}{logical: if \code{TRUE}}

\item{debug}{logical: if \code{TRUE}, echo the ffmpeg output to the console}
}
\value{
The path to the new video file, which if \code{overwrite = TRUE} will be the input file, otherwise a file in the temporary directory
}
\description{
Experimental! Requires that ffmpeg is available on your system path. Note that this function creates a new video file with the specified metadata added. This is always a file in the temporary directory. If \code{overwrite = TRUE}, the original file is deleted and replaced with the new file. Note also that the supported video tags depend on the video file type
}
\examples{
\dontrun{
  newfile <- ov_set_video_meta("blah.mp4", comment = "A comment")
  ov_get_video_meta(newfile)
}

}
\seealso{
\code{\link[=ov_get_video_meta]{ov_get_video_meta()}}
}
