% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ov_video_extract_clip}
\alias{ov_video_extract_clip}
\title{Extract clip from video file}
\usage{
ov_video_extract_clip(
  video_file,
  outfile,
  start_time,
  duration,
  end_time,
  extra = NULL,
  debug = FALSE
)
}
\arguments{
\item{video_file}{string: path to the input file}

\item{outfile}{string: path to the output file. If missing, a temporary file (with extension .mp4) will be used}

\item{start_time}{numeric: start time in seconds}

\item{duration}{numeric: duration in seconds. If missing, will be calculated from start_time and end_time}

\item{end_time}{numeric: end time in seconds. If missing, will be calculated from start_time and duration}

\item{extra}{: additional parameters passed to ffmpeg, in the form c("param", "value", "param2", "value2")}

\item{debug}{logical: if \code{TRUE}, echo the ffmpeg output to the console}
}
\value{
The path to the video clip file
}
\description{
Requires that ffmpeg is available on your system path.
}
