% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg_utils.R
\name{ov_video_frame}
\alias{ov_video_frame}
\title{Extract one or more specific frames from a video file}
\usage{
ov_video_frame(
  video_file,
  t,
  n,
  format = "jpg",
  debug = FALSE,
  framerate,
  method = "auto"
)
}
\arguments{
\item{video_file}{string: path to the video file}

\item{t}{numeric: the times of the frames to extract (in seconds)}

\item{n}{integer: the frame numbers of the frames to extract. Ignored if \code{t} is provided}

\item{format}{string: "jpg" or "png"}

\item{debug}{logical: if \code{TRUE}, echo the ffmpeg output to the console}

\item{framerate}{numeric: the framerate of the video. If not supplied, it will be found using [\link[av:info]{av::av_video_info}]}

\item{method}{string: the method to use, either "ffmpeg", "av", or "auto". "ffmpeg" is faster than "av" but requires that ffmpeg is available on your system path. If \code{method} is "auto", "ffmpeg" will be used if available and "av" if not}
}
\value{
The paths to the frame image files
}
\description{
Extract one or more specific frames from a video file
}
\examples{
video_file <- ov_example_video(1)
img <- ov_video_frame(video_file, t = 5)
img <- ov_video_frame(video_file, n = 150)
}
\seealso{
\code{\link[=ov_video_frames]{ov_video_frames()}}
}
