% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ov_video_frames}
\alias{ov_video_frames}
\title{Extract multiple frames from a video file}
\usage{
ov_video_frames(
  video_file,
  start_time,
  duration,
  end_time,
  outdir,
  fps,
  format = "jpg",
  jpg_quality = 1,
  extra = NULL,
  debug = FALSE
)
}
\arguments{
\item{video_file}{string: path to the video file}

\item{start_time}{numeric: start time in seconds}

\item{duration}{numeric: duration in seconds. If missing, will be calculated from start_time and end_time}

\item{end_time}{numeric: end time in seconds. If missing, will be calculated from start_time and duration}

\item{outdir}{string: path to the output directory, which must exist. If missing, a temporary directory will be used}

\item{fps}{numeric: frames per second, default is to extract all frames}

\item{format}{string: "jpg" or "png"}

\item{jpg_quality}{numeric: jpg quality from 1-31, lower is better (this is passed to ffmpeg as the \code{-qscale:v} parameter)}

\item{extra}{: additional parameters passed to ffmpeg, in the form c("param", "value", "param2", "value2")}

\item{debug}{logical: if \code{TRUE}, echo the ffmpeg output to the console}
}
\value{
A character vector of file names, one per frame
}
\description{
Requires that ffmpeg is available on your system path.
}
\seealso{
\code{\link{ov_video_frame}}
}
