% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{ov_video_js}
\alias{ov_video_js}
\title{Inject javascript for an HTML video player}
\usage{
ov_video_js(youtube = FALSE, version = 1)
}
\arguments{
\item{youtube}{logical: set to \code{TRUE} to include the Youtube API javascript. This isn't necessary if you are only using local video files}

\item{version}{numeric: code version. Default = 1, experimental = 2}
}
\value{
A \code{head} tag containing script tags
}
\description{
Inject javascript for an HTML video player
}
\seealso{
\code{\link{ov_video_playlist}}
}
