% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovideo-package.R
\docType{package}
\name{ovideo}
\alias{ovideo-package}
\alias{ovideo}
\title{\pkg{ovideo}}
\description{
Playlists and other video support functions from volleyball match files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openvolley/ovideo}
  \item Report bugs at \url{https://github.com/openvolley/ovideo/issues}
}

}
\author{
\strong{Maintainer}: Ben Raymond \email{ben@untan.gl}

Authors:
\itemize{
  \item Adrien Ickowicz
}

Other contributors:
\itemize{
  \item openvolley.org [originator]
}

}
