% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{ov_playlist_as_onclick}
\alias{ov_playlist_as_onclick}
\title{Convert playlist to 'onclick' string}
\usage{
ov_playlist_as_onclick(
  playlist,
  video_id,
  normalize_paths = TRUE,
  dvjs_fun = "dvjs_set_playlist_and_play",
  seamless = TRUE
)
}
\arguments{
\item{playlist}{data.frame: a playlist as returned by \code{ov_video_playlist}}

\item{video_id}{string: the id of the HTML video element to attach the playlist to}

\item{normalize_paths}{logical: if \code{TRUE}, apply \code{normalizePath} to local file paths. This will e.g. expand the tilde in paths like "~/path/to/video.mp4"}

\item{dvjs_fun}{string: the javascript function to use}

\item{seamless}{logical: if clips overlap, should we transition seamlessly from one to the next?}
}
\value{
A string suitable for inclusion as an 'onclick' tag attribute
}
\description{
Convert playlist to 'onclick' string
}
\examples{
\dontrun{
  library(shiny)

  ## hand-crafted playlist for this example
  playlist <- data.frame(video_src = "xL7qEpOdtio",
                         start_time = c(5417, 7252, 6222, 7656, 7369),
                         duration = 8,
                         type = "youtube")
  shinyApp(
      ui = fluidPage(
          ov_video_js(youtube = TRUE),
          ov_video_player(id = "yt_player", style = "height: 480px; background-color: black;"),
          tags$button("Go", onclick = ov_playlist_as_onclick(playlist, "yt_player"))
      ),
      server = function(input, output) {},
  )
}

}
