% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ov_video_frame}
\alias{ov_video_frame}
\title{Extract frame from video file}
\usage{
ov_video_frame(video_file, t)
}
\arguments{
\item{video_file}{string: path to the video file}

\item{t}{numeric: the time of the frame to extract (in seconds)}
}
\value{
The path to the jpg file
}
\description{
Requires that ffmpeg is available on your system path.
}
